requirements.txt
Scrapy==2.11.0
selenium==4.15.0
webdriver-manager==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyspark==3.5.0
pandas==2.1.3
python-dotenv>=1.0.0
scrapy-selenium==0.0.12
scrapy-user-agents==0.1.1
fake-useragent==1.4.0